refactor: Unbundle nanoarrow (#104)
It turns out that there's an incredibly high probability that downstream
projects are going to have their own nanoarrow (for testing or because
they want to do other things with arrow arrays than what geoarrow-c can
do. Because of this, the projects where it's been bundled
(geoarrow-c-geos, s2geography) usually have it twice for no good reason.
This PR removes the bundling and only builds the vendored version if the
target isn't available.