refactor: Make GeoArrowArrayReader more self-contained (#105)
The `GeoArrowArrayReader()` is the easiest point of entry for reading
"any geometry" via the visitor, even if a more efficient mechanism to
read an array might exist. To reduce the scope of the
`GeoArrowArrayView`, this PR moves the internals to be opaque with
respect to array storage so that it can do things like read unions,
dictionary-encoded, run-end encoded, or view-encoded versions of arrays
(even though this implementation won't produce those types of arrays).