feat: Support crs_type values in most recent spec release candidate (#107)
This PR add support for the `crs_type` key that was added in recent PRs
to the GeoArrow spec. We already had a `GeoArrowCrsType` enum which
simplifies things considerably; however, we'd been guessing its value
from the first character of the CRS. In this PR we stop doing that but
parse the value of `crs_type` and pass it on.

In doing this, I also improved the serialization and deserialization
code, added a canonical "lonlat" setter, and generally improved the
tests for metadata serialization/deserialization.