chore(dev/benchmarks): Benchmark coordinate sequence iteration (#112)
This PR benchmarks coordinate iteration (and fixes the previous
benchmarks benchmark the same cases). Some takeaways:

- In C, `GEOARROW_COORD_VIEW_VALUE()`, which uses `i * coord_stride`,
seems to prevent the compiler from doing some optimization which is
possible when incrementing the pointer by `coord_stride` in each
iteration of the loop. However, this only affects bounds calculation and
not the centroid calculation.
- In C++, the STL iterator does a slightly better job optimizing `i *
coord_stride` and in one case seems to do slightly better.

This PR also adds "dimension iterators" `dbegin()` and `dend()` for
iterating over a particular dimension.

My personal takeaway is that except for possibly very cheap operations
(like bounding, winding, and centroid calculations), C++ STL iteration
is probably fine. A good reason to use dimension-specific iteration is
if you are specifically aiming for autovectorization of some kind (or if
you specifically are discarding values from another dimension).