refactor: Separate out visitor-based native array writer from buffer-based writer for native arrays (#114)
This PR improves the consistency of the C API by adding the
`GeoArrowNativeWriter` to match the `GeoArrowWKTWriter` and
`GeoArrowWKBWriter`. Before, the `GeoArrowBuilder` served the purpose of
both building by buffer and building by visitor, which was slightly
confusing since it could be used to build WKB and WKT arrays too (except
when using the visitor).

This hopefully makes it more clear that the `GeoArrowArrayWriter` and
`GeoArrowArrayReader` are the things you want if you are reading
arbitrary input/generating arbitrary output.