feat: Add WKB utilities (#116)
This PR extends the iteration utilities in hpp/array_util.hpp to also
work with well-known binary, whose sequences are also interleaved
coordinates except that they are not aligned. There's some duplication
between the aligned and unaligned coordinate sequences, but the
benchmarks suggest that there's still a ~10% advantage to aligning the
coordinate sequence (at least in the somewhat synthetic benchmark
example), so I kept it in for now.

With adding WKB support, I also added in the ability to do a few
patterns of iteration (notably: `VisitVertices()` and `VisitEdges()`).
These are useful in the immediate for bounding (and testing!).

Credit to kylebarron for the overall approach in Rust Geo/WKB! (i.e.,
do a little pre-parsing to isolate the sequences to allow branchless
sequence iteration after the heavily branched parsing step).