fix(python/geoarrow-c): Update for recent Cython (#147)
This PR updates _lib such that enums are available from the module for
`cpdef enum` items. We could also refer to them directly but that would
require more code updates...all of this usage is internal and I'm not
sure how much longer this will still use Cython for, so I'm opting for
the minimal change for the moment.

Closes #145.